mysite/requirements.txt
Django==2.2.17
django-taggit==1.3.0
Markdown==3.3.3
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
